feat: move functions to tonejs engine
