+++ src/frontend/managers/MetronomeManager.js
@@ -37,13 +37,6 @@ export class MetronomeManager {
         this.visualEffectsManager = new VisualEffectsManager();
         this._audioEngine = new TonejsEngine(this);
 
-        // this._tempSequence = new Tone.Sequence((time, beat) => {
-        //    this.playStep(time, beat);
-        // }, this._beatBarsManager.beatSequence, "4n");
-
-        //Tone.getTransport().bpm.value = this.bpm * 3;
-        //this._sequence = this.generateFixedMetronomeSequence();
-
         makeAutoObservable(this)
 
     }
@@ -83,140 +76,9 @@ export class MetronomeManager {
     tempStart() {
         this._isPlaying = true;
         this._audioEngine.startPlaying();
-    }
-
-
-    getMetronomeLoopCallback() {
-        console.log("loop callback has started");
-        // this._currentStep = this._count % this._sequence.length;
-        // this._isStartOfLoop = this._currentStep === 0;
-
-        //TODO: add training mode back
-        // if (this.trainingModeManager.getIsTrainingMode()) {
-        //     if (this._isStartOfLoop &&
-        //         (this.trainingModeManager.getIsFirstLoop() || Math.random() < this.trainingModeManager.getLoopSkipProbability())) {
-        //         this._skipper = this._sequence.length;
-        //     }
-        // }
-
-        // if (this._skipper > 0) {
-        //     this._skipper--;
-        //     if (this.trainingModeManager.getIsFirstLoop()) {
-        //         this.playMetronomeStep(this._sequence, this._currentStep, time);
-        //     }
-        //     if (this._skipper === 0) {
-        //         this.trainingModeManager.setIsFirstLoop(false);
-        //     }
-        // } else {
-        //     this.playMetronomeStep(this._sequence, this._currentStep, time);
-        // }
-
-        if (this._isStartOfLoop) {
-            this._loopCount += 1;
-        }
-
-        this._count++;
-    }
-
-
-    startMetronome() {
-        this._isPlaying = true;
-        this._skipper = 0;
-        this._loop= new Tone.Loop(() => {
-            this.getMetronomeLoopCallback();
-        }, "64n");
-        console.log(Tone.getTransport().bpm.value);
-        console.log(this._sequence);
-        Tone.getTransport().start();
-        this._loop.start(0)
-        setInterval(() => {
-            console.log(Tone.getTransport().state);
-            console.log(Tone.getTransport().seconds);
-        }, 5000);
 
         //TODO: move pendulum!
         //    this.elementsManager.movePendulum();
-    }
-
-    playMetronomeStep(sequence, currentStep, time) {
-        const currentNote = sequence[currentStep];
-        if (!currentNote || !currentNote.sound) return;
-        if (!(this.trainingModeManager.getIsTrainingMode() && Math.random() < this.trainingModeManager.getNoteSkipProbability() && !this.trainingModeManager.getIsFirstLoop())) {
-            console.log("stepped here");
-            const {sound, settings} = currentNote;
-
-
-            // Динамически применяем все параметры из settings к sound
-            for (const key in settings) {
-                if (settings.hasOwnProperty(key)) {
-                    if (key in sound) {
-                        // Если параметр есть в объекте sound (например, volume)
-                        sound[key].value = settings[key];
-                    } else if (key in sound.oscillator) {
-                        // Если параметр относится к осциллятору (например, frequency, detune, phase)
-                        sound.oscillator[key] = settings[key];
-                    } else if (key in sound.envelope) {
-                        // Если параметр относится к огибающей (например, attack, decay, sustain, release)
-                        sound.envelope[key] = settings[key];
-                    } else if (key in sound.filter) {
-                        // Если параметр относится к фильтру (например, filterFrequency, filterQ, filterType)
-                        sound.filter[key] = settings[key];
-                    }
-                }
-            }
-
-            // Запускаем звук
-            sound.triggerAttackRelease('C4', '64n', time);
-
-            // Визуальные эффекты
-            elements.flashingBar.style.opacity = 1;
-            setTimeout(() => elements.flashingBar.style.opacity = 0, 100);
-
-            const beatElement = document.querySelector(`.beat[data-beat="${currentNote.beatIndex}"]`);
-            beatElement.classList.add('playing');
-            setTimeout(() => beatElement.classList.remove('playing'), 100);
-        }
-    }
-
-
-    generateFixedMetronomeSequence() {
-        let position = 0;
-        let totalSteps = 0;
-        let beatAmount = this.beatBarsManager.beats.length;
-
-
-        for (let i = 0; i < beatAmount; i++) {
-            const noteSize = this.beatBarsManager.beats[i].noteSettings.noteSize;
-            const noteAmount = this.beatBarsManager.beats[i].noteAmount;
-
-            totalSteps += 64 / noteSize * 3 * noteAmount;
-        }
-
-        const sequence = new Array(totalSteps).fill(null);
-
-        for (let beatIndex = 0; beatIndex < beatAmount; beatIndex++) {
-            const {noteSettings, noteAmount, beatSound, soundSettings} = this.beatBarsManager.beats[beatIndex];
-            const {isTriplet, noteSize} = noteSettings;
-
-            const stepSize = isTriplet ? (64 / noteSize) : (64 / noteSize * 3);
-
-
-            for (let j = 0; j < (isTriplet ? 3 * noteAmount : noteAmount); j++) {
-                sequence[position] = {beatSound, soundSettings, beatIndex};
-                position += stepSize;
-            }
-        }
-
-        return sequence;
-    }
-
-    generateMetronomeSequence() {
-
-    }
-
-    updateMetronomeSequence() {
-        this._sequence = this.generateFixedMetronomeSequence();
-        this._loop.callback = (time) => this.getMetronomeLoopCallback(time);
     }
